Fix uninitalized sensor bug
Some Ruuvi tags are missing sensors. In those cases, the sensor is of type None.
Before this commit, none type sensor values where sent.
Furthermore, the missing sensor was counted towards the number of sensor values sent
but not the number of sensor values that should be sent. This caused sensor values that
were considered late for sending not to be sent.